feat(ipc): carry broker-lane control traffic over version-checked FrameV1
When the broker lane is active and actually takes the Broker route, the
daemon control roundtrip now sends the request as a 0x7A63 FrameV1 envelope
to the broker-resolved endpoint (the daemon serve loop already auto-detects
it) instead of resolving an endpoint, dropping the negotiated session, and
re-dialing with bincode. The broker lane is version-authoritative on this
path, so there is no bincode fallback; any direct/fallback route keeps the
existing env-selected behavior byte-for-byte. Stays behind the existing
ZCCACHE_BROKER_CONNECT opt-in.

This is issue #720 Phase 1 ("carry real traffic through the broker, stop
resolve-and-drop") using the sanctioned FrameV1-lane alternative — adopting
the raw interprocess stream as the data connection (a from_io constructor)
is deferred because the negotiated stream is sync interprocess while the
zccache transport is tokio.

The ipc unit tests were split into ipc/mod_tests.rs (included via #[path])
to keep mod.rs under the repo per-file LOC ceiling.

Refs #720
